Remove TestRepository::create()
